backend/requirements.txt
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
